[UPDATE] Agregación de histogramas.

+++ DOHKO/backend/api/models.py
@@ -1,5 +1,10 @@
 from django.db import models
 import pandas as pd
+
+import matplotlib
+matplotlib.use('Agg')
+import matplotlib.pyplot as plt
+
 import numpy as np
 import os
 from django.conf import settings
@@ -10,25 +15,28 @@ class Project(models.Model):
     desc = models.CharField(max_length=200, null=True)
     url = models.CharField(max_length=200, null=True)
     dataFile = models.FileField(upload_to='datasets/', null=True, default=None)
-    
+    fileName = models.CharField(max_length=100, null=True)
 
     dataframe = None
     rows = None
     cols = None
     code = None
     dataGraph = None
+    resType = None
 
     def delete(self):
-        os.remove(os.path.join(settings.MEDIA_ROOT, self.dataFile.name))
+        if self.dataFile:
+            os.remove(os.path.join(settings.MEDIA_ROOT, self.dataFile.name))
         return super().delete()
 
     def load_data(self):
-        if not self.dataframe:
-            self.dataframe = pd.read_csv(self.dataFile or self.url)
+        self.dataframe = pd.read_csv(self.dataFile or self.url)
 
-    def getPreviewDataframe(self):
-        # self.dataframe = pd.read_csv(self.dataFile)
+
+    def dataPreview(self):
         self.load_data()
+        self.resType = 'table'
+
         # Columna para numero de fila
         self.cols = np.insert( self.dataframe.columns.values, 0, '')
 
@@ -41,29 +49,35 @@ class Project(models.Model):
         # Separacion
         self.rows.insert(5, ['...' for i in range(len(self.cols))])
 
+
     
     def dataShape(self):
         """Dimesión del dataframe."""
         self.load_data()
+        self.resType = 'text'
         self.code = {
             'command':'dataframe.shape',
-            'result':str(self.dataframe.shape)}
+            'result':str(self.dataframe.shape)
+            }
 
     def dataTypes(self):
         """Tipo de datos del dataframe."""
         self.load_data()
+        self.resType = 'table'
         self.cols = ['Variable','Tipo']
         self.rows = [(k,str(v)) for k,v in self.dataframe.dtypes.items()]
 
     def dataNull(self):
         """Datos nulos del dataframe."""
         self.load_data()
+        self.resType = 'table'
         self.cols = ['Variable','Cuenta']
         self.rows = list(self.dataframe.isnull().sum().items())
 
     def dataDescribe(self):
         """Resumen estadístico de variables numéricas."""
         self.load_data()
+        self.resType = 'table'
         self.cols = self.dataframe.describe().columns.tolist()
         self.cols.insert(0,'')
         measures = ['Cuenta', 'Media', 'Std.', 'Min','25%','50%','75%','Max']
@@ -74,30 +88,56 @@ class Project(models.Model):
 
     def dataCorrelation(self):
         self.load_data()
-        self.cols = self.dataframe.corr().columns.tolist()
-        self.rows = self.dataframe.corr().values.tolist()
-        rows_lst = self.dataframe.corr().values
-        self.rows = [ list(map(lambda cell: str(cell) if pd.isna(cell) else cell, row)) for row in rows_lst]
+        self.resType = 'table'
+        
+        self.cols = np.insert( self.dataframe.corr().columns, 0, '').tolist()
+        
+        # Redondeo a 6 digitos
+        rows_lst = [ [round(c, 6) for c in row] for row in self.dataframe.corr().values]
+        # Conversion a string + columna de variables
+        self.rows = [ [col] + [str(c) if pd.isna(c) else c for c in row] for col, row in zip(self.cols[1:], rows_lst)]
 
         data = []
-        for col, row in zip(self.cols,rows_lst):
+        for col, row in zip(self.cols, rows_lst):
             data.append(
                 {
                     'id':col,
-                    'data':[{'x':x, 'y':None if pd.isna(y) else y} for x,y in zip(self.cols,row)]
+                    'data':[ { 'x': x, 'y': None if pd.isna(y) else y } for x, y in zip( self.cols, row ) ]
                 }
             )
 
         self.dataGraph = data
 
-    def dif_values(self):
-        # Histogramas
+    def dataHistogram(self):
+        self.load_data()
+        self.resType = 'hist'
+        
+        datas = []       
+        vars = [k for k,v in self.dataframe.dtypes.items() if str(v) != 'object']
 
-        # Diagramas de bigote
+        for var in vars:
 
-        # Variables categoricas
-        self.cols = self.dataframe.describe(include='object').columns.tolist()
-        self.rows = self.dataframe.describe(include='object').values.tolist()
+            hist = self.dataframe[var].hist()
+            ax = plt.gca()
+            p = ax.patches
 
+            # Alto de las barras (y)
+            ys = [p[i].get_height() for i in range(len(p))]
+            # Inicio de las barras (x)
+            xs = [p[i].get_x() for i in range(len(p))]
+            
+            dif = (p[1].get_x() - p[0].get_x()) /2
+            xs = [ round(dif + i, 1) for i in xs]
 
+            data = [ { "id":x, "value":y } for x, y in zip(xs,ys) ]
+            
+            datas.append(
+                {
+                    'data':data,
+                    'title':var
+                }
+            )
+            plt.clf()
+
+        self.dataGraph = datas
